requirements.txt
Flask==3.0.3
firebase-admin==6.5.0
pyjwt==2.8.0
pytest==8.2.2
Flask-Testing==0.8.1
python-dotenv==1.0.1
Flask-JWT-Extended==4.6.0
mock-firestore==0.11.0
pytest-flask==1.3.0
pytest-mock==3.14.0